requirements.txt

langchain
python-dotenv
langchain-community
pypdf
chromadb

openai
flask
langchain_openai